server_requirements.txt
Django==1.6.2
MySQL-python==1.2.5
django-crispy-forms==1.4.0
django-floppyforms==1.1.1
djangorestframework==2.3.12
google-api-python-client==1.2
httplib2==0.9
psutil==1.2.1
python-dateutil==2.2
python-ldap==2.4.14
python-mimeparse==0.1.4
virtualenv==1.11.5
wsgiref==0.1.2
xlrd==0.9.2
